[SYCL][Graph] Implement exceptions for incompatible extensions
Adds info to exception message
Moves tests from e2e to unitests when possible
