+++ sycl/include/sycl/handler.hpp
@@ -381,7 +381,7 @@ private:
                             "a single kernel or explicit memory operation.");
   }
 
-  void throwIfGraphAssociated();
+  void throwIfGraphAssociated(const std::string ExceptionMsg);
 
   constexpr static int AccessTargetMask = 0x7ff;
   /// According to section 4.7.6.11. of the SYCL specification, a local accessor
@@ -2088,10 +2088,9 @@ public:
   std::enable_if_t<
       ext::oneapi::experimental::is_property_list<PropertiesT>::value>
   single_task(PropertiesT Props, _KERNELFUNCPARAM(KernelFunc)) {
-    if (!std::is_same<
-            PropertiesT,
-            ext::oneapi::experimental::detail::empty_properties_t>::value)
-      throwIfGraphAssociated();
+    if (!std::is_same_v<PropertiesT,
+                        ext::oneapi::experimental::detail::empty_properties_t>)
+      throwIfGraphAssociated("sycl_ext_oneapi_kernel_properties");
     single_task_lambda_impl<KernelName, KernelType, PropertiesT>(Props,
                                                                  KernelFunc);
   }
@@ -2102,10 +2101,9 @@ public:
       ext::oneapi::experimental::is_property_list<PropertiesT>::value>
   parallel_for(range<1> NumWorkItems, PropertiesT Props,
                _KERNELFUNCPARAM(KernelFunc)) {
-    if (!std::is_same<
-            PropertiesT,
-            ext::oneapi::experimental::detail::empty_properties_t>::value)
-      throwIfGraphAssociated();
+    if (!std::is_same_v<PropertiesT,
+                        ext::oneapi::experimental::detail::empty_properties_t>)
+      throwIfGraphAssociated("sycl_ext_oneapi_kernel_properties");
     parallel_for_lambda_impl<KernelName, KernelType, 1, PropertiesT>(
         NumWorkItems, Props, std::move(KernelFunc));
   }
@@ -2116,10 +2114,9 @@ public:
       ext::oneapi::experimental::is_property_list<PropertiesT>::value>
   parallel_for(range<2> NumWorkItems, PropertiesT Props,
                _KERNELFUNCPARAM(KernelFunc)) {
-    if (!std::is_same<
-            PropertiesT,
-            ext::oneapi::experimental::detail::empty_properties_t>::value)
-      throwIfGraphAssociated();
+    if (!std::is_same_v<PropertiesT,
+                        ext::oneapi::experimental::detail::empty_properties_t>)
+      throwIfGraphAssociated("sycl_ext_oneapi_kernel_properties");
     parallel_for_lambda_impl<KernelName, KernelType, 2, PropertiesT>(
         NumWorkItems, Props, std::move(KernelFunc));
   }
@@ -2130,10 +2127,9 @@ public:
       ext::oneapi::experimental::is_property_list<PropertiesT>::value>
   parallel_for(range<3> NumWorkItems, PropertiesT Props,
                _KERNELFUNCPARAM(KernelFunc)) {
-    if (!std::is_same<
-            PropertiesT,
-            ext::oneapi::experimental::detail::empty_properties_t>::value)
-      throwIfGraphAssociated();
+    if (!std::is_same_v<PropertiesT,
+                        ext::oneapi::experimental::detail::empty_properties_t>)
+      throwIfGraphAssociated("sycl_ext_oneapi_kernel_properties");
     parallel_for_lambda_impl<KernelName, KernelType, 3, PropertiesT>(
         NumWorkItems, Props, std::move(KernelFunc));
   }
@@ -2144,10 +2140,9 @@ public:
       ext::oneapi::experimental::is_property_list<PropertiesT>::value>
   parallel_for(nd_range<Dims> Range, PropertiesT Properties,
                _KERNELFUNCPARAM(KernelFunc)) {
-    if (!std::is_same<
-            PropertiesT,
-            ext::oneapi::experimental::detail::empty_properties_t>::value)
-      throwIfGraphAssociated();
+    if (!std::is_same_v<PropertiesT,
+                        ext::oneapi::experimental::detail::empty_properties_t>)
+      throwIfGraphAssociated("sycl_ext_oneapi_kernel_properties");
     parallel_for_impl<KernelName>(Range, Properties, std::move(KernelFunc));
   }
 
@@ -2160,10 +2155,9 @@ public:
       detail::AreAllButLastReductions<RestT...>::value &&
       ext::oneapi::experimental::is_property_list<PropertiesT>::value>
   parallel_for(range<1> Range, PropertiesT Properties, RestT &&...Rest) {
-    if (!std::is_same<
-            PropertiesT,
-            ext::oneapi::experimental::detail::empty_properties_t>::value)
-      throwIfGraphAssociated();
+    if (!std::is_same_v<PropertiesT,
+                        ext::oneapi::experimental::detail::empty_properties_t>)
+      throwIfGraphAssociated("sycl_ext_oneapi_kernel_properties");
     detail::reduction_parallel_for<KernelName>(*this, Range, Properties,
                                                std::forward<RestT>(Rest)...);
   }
@@ -2175,10 +2169,9 @@ public:
       detail::AreAllButLastReductions<RestT...>::value &&
       ext::oneapi::experimental::is_property_list<PropertiesT>::value>
   parallel_for(range<2> Range, PropertiesT Properties, RestT &&...Rest) {
-    if (!std::is_same<
-            PropertiesT,
-            ext::oneapi::experimental::detail::empty_properties_t>::value)
-      throwIfGraphAssociated();
+    if (!std::is_same_v<PropertiesT,
+                        ext::oneapi::experimental::detail::empty_properties_t>)
+      throwIfGraphAssociated("sycl_ext_oneapi_kernel_properties");
     detail::reduction_parallel_for<KernelName>(*this, Range, Properties,
                                                std::forward<RestT>(Rest)...);
   }
@@ -2190,10 +2183,9 @@ public:
       detail::AreAllButLastReductions<RestT...>::value &&
       ext::oneapi::experimental::is_property_list<PropertiesT>::value>
   parallel_for(range<3> Range, PropertiesT Properties, RestT &&...Rest) {
-    if (!std::is_same<
-            PropertiesT,
-            ext::oneapi::experimental::detail::empty_properties_t>::value)
-      throwIfGraphAssociated();
+    if (!std::is_same_v<PropertiesT,
+                        ext::oneapi::experimental::detail::empty_properties_t>)
+      throwIfGraphAssociated("sycl_ext_oneapi_kernel_properties");
     detail::reduction_parallel_for<KernelName>(*this, Range, Properties,
                                                std::forward<RestT>(Rest)...);
   }
@@ -2248,10 +2240,9 @@ public:
             int Dims, typename PropertiesT>
   void parallel_for_work_group(range<Dims> NumWorkGroups, PropertiesT Props,
                                _KERNELFUNCPARAM(KernelFunc)) {
-    if (!std::is_same<
-            PropertiesT,
-            ext::oneapi::experimental::detail::empty_properties_t>::value)
-      throwIfGraphAssociated();
+    if (!std::is_same_v<PropertiesT,
+                        ext::oneapi::experimental::detail::empty_properties_t>)
+      throwIfGraphAssociated("sycl_ext_oneapi_kernel_properties");
     parallel_for_work_group_lambda_impl<KernelName, KernelType, Dims,
                                         PropertiesT>(NumWorkGroups, Props,
                                                      KernelFunc);
@@ -2262,10 +2253,9 @@ public:
   void parallel_for_work_group(range<Dims> NumWorkGroups,
                                range<Dims> WorkGroupSize, PropertiesT Props,
                                _KERNELFUNCPARAM(KernelFunc)) {
-    if (!std::is_same<
-            PropertiesT,
-            ext::oneapi::experimental::detail::empty_properties_t>::value)
-      throwIfGraphAssociated();
+    if (!std::is_same_v<PropertiesT,
+                        ext::oneapi::experimental::detail::empty_properties_t>)
+      throwIfGraphAssociated("sycl_ext_oneapi_kernel_properties");
     parallel_for_work_group_lambda_impl<KernelName, KernelType, Dims,
                                         PropertiesT>(
         NumWorkGroups, WorkGroupSize, Props, KernelFunc);
@@ -2573,7 +2563,7 @@ public:
   /// until all commands previously submitted to this queue have entered the
   /// complete state.
   void ext_oneapi_barrier() {
-    throwIfGraphAssociated();
+    throwIfGraphAssociated("sycl_ext_oneapi_enqueue_barrier");
     throwIfActionIsCreated();
     setType(detail::CG::Barrier);
   }
@@ -2659,7 +2649,7 @@ public:
             typename = std::enable_if_t<std::is_same_v<T, unsigned char>>>
   void ext_oneapi_memcpy2d(void *Dest, size_t DestPitch, const void *Src,
                            size_t SrcPitch, size_t Width, size_t Height) {
-    throwIfGraphAssociated();
+    throwIfGraphAssociated("sycl_ext_oneapi_memcpy2d");
     throwIfActionIsCreated();
     if (Width > DestPitch)
       throw sycl::exception(sycl::make_error_code(errc::invalid),
@@ -2838,7 +2828,7 @@ public:
   void memcpy(ext::oneapi::experimental::device_global<T, PropertyListT> &Dest,
               const void *Src, size_t NumBytes = sizeof(T),
               size_t DestOffset = 0) {
-    throwIfGraphAssociated();
+    throwIfGraphAssociated("sycl_ext_oneapi_device_global");
     if (sizeof(T) < DestOffset + NumBytes)
       throw sycl::exception(make_error_code(errc::invalid),
                             "Copy to device_global is out of bounds.");
@@ -2871,7 +2861,7 @@ public:
   memcpy(void *Dest,
          const ext::oneapi::experimental::device_global<T, PropertyListT> &Src,
          size_t NumBytes = sizeof(T), size_t SrcOffset = 0) {
-    throwIfGraphAssociated();
+    throwIfGraphAssociated("sycl_ext_oneapi_device_global");
     if (sizeof(T) < SrcOffset + NumBytes)
       throw sycl::exception(make_error_code(errc::invalid),
                             "Copy from device_global is out of bounds.");
