+++ sycl/test-e2e/Graph/graph_exception_global_device_extension.cpp
@@ -1,3 +1,4 @@
+// REQUIRES: level_zero, gpu
 // RUN: %{build} -o %t.out
 // RUN: %{run} %t.out
 //
